project-1-aviation-multi-agent/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
